Add RetryPolicy Removes IRetry (#122)
Add RetryPolicy Removes IRetry

This change adds the RetryPolicy type which represents the retry policy to use as an union type of the different supported retry policies.
This is an initial change that only replaces RetryPolicy with IRetry and keeps the same current behavior. However, this change will enable us to store the retry policy with the DurablePromise to be used in the recovery path.