Add handleCompletePromise helper function
We had some code repetition around handling a completed durable promise.
The repetition was fine, how ever this pattern will come to callbacks
and tasks and we will have it in atleast 4 more places, having to
write that code that many times is error prone and verbose.

This change also fixes a bug made evident by this helper funciton where
we were returning the value computed by the function directly instead of
the value stored in the completed durable promise.