Fix sleep tests.
The current sleep tests had timeouts set too short, given the
server <-> client latencies and the different processing times
both at the server and client it was needed to increase the timeouts
and sleep times. This caused the tests to take way too long, by merging
all the test calls in a single one and awaiting only a "promise.all"
it was possible to speed it up.